Added Postgres to the platform as a new chart.
It provides a statefulset which allows HA and
data persistency for the data stored in Postgres.

The chart contains a secret which stores the
auth and db data.

It also contains a NodePort service which will
allow us to use it using <node-ip>:<node-port>.

The stateful set deploys 2 replicas of Postgres,
each of them will have a PVC automatically
provisioned by the local-path StorageClass.

Todo:
- create an ingress to allow persistent
trafic from outside the cluster.